requirements.txt
Django
djangorestframework
sqlite3